+++ tests/lib/Unit/Store/StoreTest.php
@@ -22,7 +22,7 @@ use CloudCreativity\LaravelJsonApi\Contracts\Adapter\RelationshipAdapterInterfac
 use CloudCreativity\LaravelJsonApi\Contracts\Adapter\ResourceAdapterInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\ContainerInterface;
 use CloudCreativity\LaravelJsonApi\Contracts\Store\StoreInterface;
-use CloudCreativity\LaravelJsonApi\Exceptions\RecordNotFoundException;
+use CloudCreativity\LaravelJsonApi\Exceptions\ResourceNotFoundException;
 use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use CloudCreativity\LaravelJsonApi\Object\ResourceIdentifier;
 use CloudCreativity\LaravelJsonApi\Object\ResourceIdentifierCollection;
@@ -219,38 +219,33 @@ class StoreTest extends TestCase
 
     public function testExistsWithIdentifier()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $store = $this->store([
             'posts' => $this->adapter(),
             'users' => $this->willExist('99')
         ]);
 
         $this->assertTrue($store->isType('users'));
-        $this->assertTrue($store->exists($identifier));
+        $this->assertTrue($store->exists('users', '99'));
     }
 
     public function testDoesNotExist()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $store = $this->store([
             'posts' => $this->adapter(),
             'users' => $this->willNotExist('99')
         ]);
 
         $this->assertTrue($store->isType('users'));
-        $this->assertFalse($store->exists($identifier));
+        $this->assertFalse($store->exists('users', '99'));
     }
 
     public function testCannotDetermineExistence()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
         $store = $this->store(['posts' => $this->adapter()]);
 
         $this->assertFalse($store->isType('users'));
         $this->expectException(RuntimeException::class);
-        $store->exists($identifier);
+        $store->exists('users', '99');
     }
 
     public function testFind()
@@ -267,7 +262,6 @@ class StoreTest extends TestCase
 
     public function testFindWithIdentifier()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
         $expected = new StandardObject();
 
         $store = $this->store([
@@ -275,23 +269,21 @@ class StoreTest extends TestCase
             'users' => $this->willFind('99', $expected)
         ]);
 
-        $this->assertSame($expected, $store->find($identifier));
-        $this->assertSame($expected, $store->findOrFail($identifier));
+        $this->assertSame($expected, $store->find('users', '99'));
+        $this->assertSame($expected, $store->findOrFail('users', '99'));
     }
 
     public function testCannotFind()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $store = $this->store([
             'posts' => $this->adapter(),
             'users' => $this->willNotFind('99')
         ]);
 
-        $this->assertNull($store->find($identifier));
-        $this->expectException(RecordNotFoundException::class);
-        $this->expectExceptionMessage('users:99');
-        $store->findOrFail($identifier);
+        $this->assertNull($store->find('users', '99'));
+        $this->expectException(ResourceNotFoundException::class);
+        $this->expectExceptionMessage("Resource users with id 99 does not exist.");
+        $store->findOrFail('users', '99');
     }
 
     /**
@@ -299,15 +291,13 @@ class StoreTest extends TestCase
      */
     public function testExistsCalledOnce()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $store = $this->store([
             'posts' => $this->adapter(),
             'users' => $this->willExist('99', true, $this->once())
         ]);
 
-        $this->assertTrue($store->exists($identifier));
-        $this->assertTrue($store->exists($identifier));
+        $this->assertTrue($store->exists('users', '99'));
+        $this->assertTrue($store->exists('users', '99'));
     }
 
     /**
@@ -315,7 +305,6 @@ class StoreTest extends TestCase
      */
     public function testFindCalledOnce()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
         $expected = new StandardObject();
 
         $store = $this->store([
@@ -323,8 +312,8 @@ class StoreTest extends TestCase
             'users' => $this->willFind('99', $expected, $this->once()),
         ]);
 
-        $this->assertSame($expected, $store->find($identifier));
-        $this->assertSame($expected, $store->find($identifier));
+        $this->assertSame($expected, $store->find('users', '99'));
+        $this->assertSame($expected, $store->find('users', '99'));
     }
 
     /**
@@ -333,7 +322,6 @@ class StoreTest extends TestCase
      */
     public function testFindBeforeExists()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
         $expected = new StandardObject();
 
         $mock = $this->adapter();
@@ -341,8 +329,8 @@ class StoreTest extends TestCase
         $mock->method('find')->with('99')->willReturn($expected);
 
         $store = $this->store(['users' => $mock]);
-        $this->assertSame($expected, $store->find($identifier));
-        $this->assertTrue($store->exists($identifier));
+        $this->assertSame($expected, $store->find('users', '99'));
+        $this->assertTrue($store->exists('users', '99'));
     }
 
     /**
@@ -351,14 +339,12 @@ class StoreTest extends TestCase
      */
     public function testFindNoneBeforeExists()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $mock = $this->adapter();
         $mock->expects($this->never())->method('exists');
 
         $store = $this->store(['users' => $mock]);
-        $this->assertNull($store->find($identifier));
-        $this->assertFalse($store->exists($identifier));
+        $this->assertNull($store->find('users', '99'));
+        $this->assertFalse($store->exists('users', '99'));
     }
 
     /**
@@ -367,15 +353,13 @@ class StoreTest extends TestCase
      */
     public function testDoesNotExistBeforeFind()
     {
-        $identifier = ResourceIdentifier::create('users', '99');
-
         $mock = $this->adapter();
         $mock->expects($this->once())->method('exists')->with('99')->willReturn(false);
         $mock->expects($this->never())->method('find');
 
         $store = $this->store(['users' => $mock]);
-        $this->assertFalse($store->exists($identifier));
-        $this->assertNull($store->find($identifier));
+        $this->assertFalse($store->exists('users', '99'));
+        $this->assertNull($store->find('users', '99'));
     }
 
     /**
@@ -384,11 +368,11 @@ class StoreTest extends TestCase
      */
     public function testFindManyReturnsEmpty()
     {
-        $identifiers = ResourceIdentifierCollection::create([
-            (object) ['type' => 'posts', 'id' => '1'],
-            (object) ['type' => 'users', 'id' => '99'],
-            (object) ['type' => 'posts', 'id' => '3'],
-        ]);
+        $identifiers = [
+            ['type' => 'posts', 'id' => '1'],
+            ['type' => 'users', 'id' => '99'],
+            ['type' => 'posts', 'id' => '3'],
+        ];
 
         $store = $this->store([
             'posts' => $this->willFindMany(['1', '3']),
@@ -429,11 +413,14 @@ class StoreTest extends TestCase
      */
     public function testFindManyWithIdentifiers()
     {
-        $identifiers = ResourceIdentifierCollection::create([
-            $post = (object) ['type' => 'posts', 'id' => '1'],
-            (object) ['type' => 'posts', 'id' => '3'],
-            $user = (object) ['type' => 'users', 'id' => '99'],
-        ]);
+        $post = (object) ['type' => 'posts', 'id' => '1'];
+        $user = (object) ['type' => 'users', 'id' => '99'];
+
+        $identifiers = [
+            ['type' => 'posts', 'id' => '1'],
+            ['type' => 'posts', 'id' => '3'],
+            ['type' => 'users', 'id' => '99'],
+        ];
 
         $store = $this->store([
             'posts' => $this->willFindMany(['1', '3'], [$post]),
@@ -450,11 +437,11 @@ class StoreTest extends TestCase
      */
     public function testCannotFindMany()
     {
-        $identifiers = ResourceIdentifierCollection::create([
-            $post = (object) ['type' => 'posts', 'id' => '1'],
-            (object) ['type' => 'posts', 'id' => '3'],
-            $user = (object) ['type' => 'users', 'id' => '99'],
-        ]);
+        $identifiers = [
+            ['type' => 'posts', 'id' => '1'],
+            ['type' => 'posts', 'id' => '3'],
+            ['type' => 'users', 'id' => '99'],
+        ];
 
         $store = $this->store([
             'posts' => $this->willFindMany(['1', '3']),
@@ -567,7 +554,7 @@ class StoreTest extends TestCase
         $mock = $this->adapter();
         $mock->expects($this->atLeastOnce())
             ->method('findMany')
-            ->with($resourceIds)
+            ->with(collect($resourceIds))
             ->willReturn($results);
 
         return $mock;